chore: change the file extensions of jsx files
