+++ majorfolio/src/stories/AllDivider.stories.jsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import AllDivider from '../components/AllDivider/AllDivider';
+
+export default {
+  title: 'All/AllDivider',
+  component: AllDivider,
+  argTypes: {},
+};
+
+function Template(args) {
+  return <AllDivider {...args} />;
+}
+
+export const Divider = Template.bind({});
+
+Divider.args = {};
