--- majorfolio/src/stories/AllTagSmall.stories.js
@@ -1,17 +1,0 @@
-import AllTagSmall from "../components/AllTagSmall/AllTagSmall";
-
-export default {
-  title: "All/AllTagSmall",
-  component: AllTagSmall,
-  argTypes: {
-    text: { control: "text" },
-  },
-};
-
-const Template = (args) => <AllTagSmall {...args} />;
-
-export const GrayTag = Template.bind({});
-
-GrayTag.args = {
-  text: "PDF",
-};
